+++ agent/wpthook/results.cc
@@ -946,45 +946,44 @@ CStringA Results::FormatTime(LARGE_INTEGER t) {
   HTML will still be captured
 -----------------------------------------------------------------------------*/
 void Results::SaveResponseBodies(void) {
-  CString file = _file_base + _T("_bodies.zip");
-  zipFile zip = zipOpen(CT2A(file), APPEND_STATUS_CREATE);
-  if (zip) {
-    DWORD count = 0;
-    DWORD bodies_count = 0;
-    _requests.Lock();
-    bool saved = false;
-    POSITION pos = _requests._requests.GetHeadPosition();
-    while (pos) {
-      Request * request = _requests._requests.GetNext(pos);
-      if (request && request->_processed) {
-        CString mime = request->GetResponseHeader("content-type").MakeLower();
-        count++;
-        if (request->GetResult() == 200 && 
-            (_test._save_response_bodies || !saved) &&
-            ( mime.Find(_T("text/")) >= 0 || 
-              mime.Find(_T("javascript")) >= 0 || 
-              mime.Find(_T("json")) >= 0))  {
-          DataChunk body = request->_response_data.GetBody(true);
-          LPBYTE body_data = (LPBYTE)body.GetData();
-          DWORD body_len = body.GetLength();
-          if (body_data && body_len) {
-            saved = true;
-            CStringA name;
-            name.Format("%03d-response.txt", count);
-            if (!zipOpenNewFileInZip(zip, name, 0, 0, 0, 0, 0, 0, Z_DEFLATED, 
-                Z_BEST_COMPRESSION)) {
-              zipWriteInFileInZip(zip, body_data, body_len);
-              zipCloseFileInZip(zip);
-              bodies_count++;
+  if (_test._save_response_bodies) {
+    CString file = _file_base + _T("_bodies.zip");
+    zipFile zip = zipOpen(CT2A(file), APPEND_STATUS_CREATE);
+    if (zip) {
+      DWORD count = 0;
+      DWORD bodies_count = 0;
+      _requests.Lock();
+      POSITION pos = _requests._requests.GetHeadPosition();
+      while (pos) {
+        Request * request = _requests._requests.GetNext(pos);
+        if (request && request->_processed) {
+          CString mime = request->GetResponseHeader("content-type").MakeLower();
+          count++;
+          if (request->GetResult() == 200 && 
+              ( mime.Find(_T("text/")) >= 0 || 
+                mime.Find(_T("javascript")) >= 0 || 
+                mime.Find(_T("json")) >= 0))  {
+            DataChunk body = request->_response_data.GetBody(true);
+            LPBYTE body_data = (LPBYTE)body.GetData();
+            DWORD body_len = body.GetLength();
+            if (body_data && body_len) {
+              CStringA name;
+              name.Format("%03d-response.txt", count);
+              if (!zipOpenNewFileInZip(zip, name, 0, 0, 0, 0, 0, 0, Z_DEFLATED, 
+                  Z_BEST_COMPRESSION)) {
+                zipWriteInFileInZip(zip, body_data, body_len);
+                zipCloseFileInZip(zip);
+                bodies_count++;
+              }
             }
           }
         }
       }
+      _requests.Unlock();
+      zipClose(zip, 0);
+      if(!bodies_count)
+        DeleteFile(file);
     }
-    _requests.Unlock();
-    zipClose(zip, 0);
-    if(!bodies_count)
-      DeleteFile(file);
   }
 }
